refactor: Move ModuleTestDebugLogging to support package
